% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnchoredFusion.R
\name{runAnchoredFusion}
\alias{runAnchoredFusion}
\title{runAnchoredFusion}
\usage{
runAnchoredFusion(step = c("bam-consolidate", "breakpoint-consolidate",
  "breakpoint-filter", "breakpoint-anno", "breakpoint-anno-post"), runInfo,
  output, sample.id)
}
\arguments{
\item{step}{step name in which you want to run AnchoredFusion, included "bam-consolidate",
"breakpoint-consolidate", "breakpoint-filter", "breakpoint-anno" and "breakpoint-anno-post"
five steps.}

\item{runInfo}{config file which sets the path and parameters of depended tools.}

\item{output}{directory of output.}

\item{sample.id}{sample name (matches with prefixed name of raw fastq or bam file).}
}
\description{
Main function of AnchoredFusion software.
}
\details{
(https://github.com/Zheng-NGS-Lab/AnchoredFusion/blob/master/README.md).

A fast pipeline for detection of gene fusion based on fusion-supporting split alignment.
}
\examples{

runSplitFuion(step = c("bam-consolidate", "breakpoint-consolidate", "breakpoint-filter"),
runInfo = "/path/example.runInfo", output = "/path/result", sample.id = "example")


}
