% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SplitFusion.breakpoint.anno.postscript.R
\name{SplitFusion.breakpoint.anno.postscript}
\alias{SplitFusion.breakpoint.anno.postscript}
\title{SplitFusion.breakpoint.anno.postscript}
\usage{
SplitFusion.breakpoint.anno.postscript(runInfo)
}
\arguments{
\item{runInfo}{config file which sets the path and parameters of depended tools.}
}
\description{
Post annotation of breakpoint detected by SplitFusion.
}
\examples{

#"anno.left.right" file outputted by last step "breakpoint-anno" of SplitFusion should be exist.

SplitFusion.breakpoint.anno.postscript(runInfo = "/path/example.runInfo")

}
